@@ -1,5 +1,5 @@
 module github.com/patrickmn/go-cache
 
-go 1.21.0
+go 1.22.0
 
 require go.uber.org/goleak v1.2.1
